CTPA-1098 xsd change

+++ src/main/java/uk/gov/ons/ctp/response/notify/endpoint/ManualTestEndpoint.java
@@ -33,7 +33,7 @@ public class ManualTestEndpoint implements CTPEndpoint {
   private static final BigDecimal LATITUDE = new BigDecimal("1000.00");
   private static final BigDecimal LONGITUDE = new BigDecimal("1000.00");
 
-  private static final BigInteger CASEID = new BigInteger("1");
+  private static final Integer CASEID = new Integer("1");
   private static final BigInteger UPRN = new BigInteger("201");
 
   private static final String ACTION_PLAN = "abc";
